Move parse_source_files from conda_lock.py to src_parser/__init__.py
